Update framework (#4) (#5)
* Ready to Improve Engine

TO DO:

Strip the context manaement part and make it thread and multi-context aware.

* Threading added

Trying to implement a thread pool

* ThreadPool class finished

Finished writing a thread-safe thread pool class. Starting to learn network IO with sockets.

* Starting to write TCP server and client

* Starting to find solution with less locks

* Removed some mutex lock uses

Removed the heavy overhead of mutex locking when polling , pushing and popping in the thread pool.

But also broke many other parts and need much fixing.

* ThreadPool class update

Removed most of the needs of a mutex for light-weight tasks that will be called multiple times.

* ThreadPool Update

* Fixed deadlock with queue with std::this_thread::yield()

Thanks to @James on Stack Overflow on the suggestion of using intrinsic function for waiting.

* Improve Threading

Removed redundant functions, added checks and added other comments.

* Start remaking the queue

Start to reimplement the queue to not locking if it does not need reallocation.

* failed...

Failed to inprove queue to have no spin-lock.

* Semester started, stashing all...

* Update environment

* Track dependencies with git-lfs

* Project directory restructure

* Remove unused library

* Add some references to a file

* Fixes with changed directory heiarchy

* Update README.md

* Update README.md

* Create LICENSE-MIT

* Update LICENSE-GPLv3

* Create LICENSE

* Simple state updated

* Add compiled samples

* Ready to Improve Engine

TO DO:

Strip the context manaement part and make it thread and multi-context aware.

* Threading added

Trying to implement a thread pool

* ThreadPool class finished

Finished writing a thread-safe thread pool class. Starting to learn network IO with sockets.

* Starting to write TCP server and client

* Starting to find solution with less locks

* Removed some mutex lock uses

Removed the heavy overhead of mutex locking when polling , pushing and popping in the thread pool.

But also broke many other parts and need much fixing.

* ThreadPool class update

Removed most of the needs of a mutex for light-weight tasks that will be called multiple times.

* ThreadPool Update

* Fixed deadlock with queue with std::this_thread::yield()

Thanks to @James on Stack Overflow on the suggestion of using intrinsic function for waiting.

* Improve Threading

Removed redundant functions, added checks and added other comments.

* Start remaking the queue

Start to reimplement the queue to not locking if it does not need reallocation.

* failed...

Failed to inprove queue to have no spin-lock.

* Semester started, stashing all...

* Update environment

* Track dependencies with git-lfs

* Project directory restructure

* Remove unused library

* Add some references to a file

* Fixes with changed directory heiarchy

* Simple state updated

* Update README.md

* Update README.md

* Add compiled samples

* Create LICENSE-MIT

* Update LICENSE-GPLv3

* Create LICENSE